+++ app/src/main/res/layout/activity_splash_screen.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.systers.mentorship.view.ui.SplashScreenActivity">
+    tools:context="org.systers.mentorship.view.activities.SplashScreenActivity">
 
     <ImageView
         android:layout_width="wrap_content"
